Move back to `rizzcomic`
