+++ src/en/rizzcomic/src/eu/kanade/tachiyomi/extension/en/rizzcomic/Filters.kt
@@ -1,4 +1,4 @@
-package eu.kanade.tachiyomi.extension.en.realmoasis
+package eu.kanade.tachiyomi.extension.en.rizzcomic
 
 import eu.kanade.tachiyomi.source.model.Filter
 import eu.kanade.tachiyomi.source.model.FilterList
